requirements.txt
flask>=2.3.0
streamlit>=1.28.0
werkzeug>=2.3.0
pillow>=10.0.0
lxml>=4.9.0
